PASS.conf.ascがなかった場合に新規作成する処理を追加

+++ quest/linux/shellscript/Password_Manager2.sh
@@ -6,8 +6,12 @@ echo 'パスワードマネージャーへようこそ!'
 read -p "次の選択肢から入力してください(Add Password/Get Password/Exit):" input
 
 if [ "$input" = "Add Password" ]; then
-	gpg PASS.conf.asc 2> /dev/null
-	rm PASS.conf.asc
+	if   [ -e PASS.conf.asc ]; then
+        	gpg PASS.conf.asc > /dev/null && rm PASS.conf.asc
+	else
+		touch PASS.conf
+		echo "PASS.confを新規作成しました"
+	fi
         read -p "サービス名を入力してください:" ServiceName
         read -p "ユーザー名を入力してください:" UserName
         read -p "パスワードを入力してください:" PassWord
@@ -15,26 +19,29 @@ if [ "$input" = "Add Password" ]; then
         gpg -e -a -r "$mail_address" PASS.conf
 	rm PASS.conf
 elif [ "$input" = "Get Password" ]; then
-	gpg PASS.conf.asc 2> /dev/null
-	rm PASS.conf.asc
-        read -p "サービス名を入力してください:" ServiceName 
-        while IFS=';' read -r -a elements; do
-		if [ "${elements[0]}" = "$ServiceName" ]; then
-	    		ServiceName1="${elements[0]}"
-	    		   UserName1="${elements[1]}"
-	    		   PassWord1="${elements[2]}"
-		fi
-	done < PASS.conf
+	if [ -e PASS.conf.asc ]; then
+		gpg PASS.conf.asc > /dev/null && rm PASS.conf.asc
+        	read -p "サービス名を入力してください:" ServiceName 
+        	while IFS=';' read -r -a elements; do
+			if [ "${elements[0]}" = "$ServiceName" ]; then
+	    			ServiceName1="${elements[0]}"
+	    		   	UserName1="${elements[1]}"
+	    		   	PassWord1="${elements[2]}"
+			fi
+		done < PASS.conf
 
-        if [ -n "$ServiceName1" ]; then 
-		echo "$ServiceName1"
-		echo "$UserName1"
-        	echo "$PassWord1"
+        	if [ -n "$ServiceName1" ]; then 
+			echo "$ServiceName1"
+			echo "$UserName1"
+        		echo "$PassWord1"
+		else
+			echo "そのサービスは登録されていません"
+		fi
+		gpg -e -a -r "$mail_address" PASS.conf
+		rm PASS.conf
 	else
-		echo "そのサービスは登録されていません"
+	        echo "PASS.confがありません"
 	fi
-	gpg -e -a -r "$mail_address" PASS.conf
-	rm PASS.conf
 elif [ "$input" = "Exit" ]; then
 	echo "Thank you!"
 else
